# docs/openapi.yaml
openapi: 3.1.0
info:
  title: MisVales API
  version: 1.0.0
paths:
...
  /api/v1/distributors/{distributor}/credit-increase-requests:
    post:
      operationId: createCreditIncreaseRequest
      summary: Registra de forma idempotente una solicitud propia
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/DistributorId' }
        - { $ref: '#/components/parameters/IdempotencyKey' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateCreditIncreaseInput' }
            examples:
              normal:
                value:
                  requested_amount: '10000.00'
                  reason: Incremento solicitado para continuar operaciones
              insufficientCredit:
                value:
                  requested_amount: '5000.00'
                  reason: El producto seleccionado supera el saldo disponible
                  origin: { type: INSUFFICIENT_CREDIT, product_amount: '15000.00' }
      responses:
        '201':
          description: Solicitud registrada o resultado idempotente existente
          content:
            application/json:
              schema: { type: object, properties: { data: { $ref: '#/components/schemas/CreditIncreaseRequest' } } }
        '403': { $ref: '#/components/responses/AuthScopeDenied' }
        '404': { $ref: '#/components/responses/CreditLineNotFound' }
        '422': { description: Importe, motivo u origen inválido }
...
  /api/v1/credit-increase-requests/{creditIncreaseRequest}/preauthorize:
    post:
      operationId: reviewCreditIncreaseRequest
      summary: Preautoriza o rechaza como coordinador responsable
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/CreditIncreaseRequestId' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CoordinatorCreditDecisionInput' }
            examples:
              preauthorize:
                value: { decision: PREAUTHORIZE, recommended_amount: '8000.00', reason: Historial y comportamiento revisados, lock_version: 1 }
              reject:
                value: { decision: REJECT, reason: La solicitud no fue preautorizada, lock_version: 1 }
      responses:
        '200': { description: Solicitud revisada }
        '403': { $ref: '#/components/responses/AuthScopeDenied' }
        '409': { description: Estado o versión incompatible }
        '422': { description: Decisión o importe inválido }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer }
  parameters:
...
    DistributorId:
      name: distributor
      in: path
      required: true
      schema: { type: string, format: uuid }
    CreditIncreaseRequestId:
      name: creditIncreaseRequest
      in: path
      required: true
      schema: { type: string, format: uuid }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 8, maxLength: 160 }
  responses:
    AuthScopeDenied:
      description: Actor sin rol, sucursal, asignación o propiedad
      content:
        application/json:
          schema: { $ref: '#/components/schemas/DomainError' }
    CreditLineNotFound:
      description: No existe una línea para la distribuidora
      content:
        application/json:
          schema: { $ref: '#/components/schemas/DomainError' }
...
  schemas:
    Money:
      type: string
      pattern: '^\d+\.\d{2}$'
      example: '20000.00'
      description: Importe decimal serializado como cadena
...
    CreditIncreaseStatus:
      type: string
      enum: [SOLICITADO, PREAUTORIZADO, RECHAZADO_COORDINADOR, RECHAZADO_GERENTE, AUTORIZADO_TOTAL, AUTORIZADO_PARCIAL, RESTRICCION_50_ACTIVA, COMPLETADO]
...
    CreateCreditIncreaseInput:
      type: object
      required: [requested_amount, reason]
      properties:
        requested_amount: { $ref: '#/components/schemas/Money' }
        reason: { type: string, minLength: 5, maxLength: 2000 }
        origin:
          type: object
          required: [type]
          properties:
            type: { type: string, enum: [NORMAL, INSUFFICIENT_CREDIT] }
            product_amount: { $ref: '#/components/schemas/Money' }
    CoordinatorCreditDecisionInput:
      type: object
      required: [decision, reason]
      properties:
        decision: { type: string, enum: [PREAUTHORIZE, REJECT] }
        recommended_amount: { $ref: '#/components/schemas/Money' }
        reason: { type: string, minLength: 5, maxLength: 2000 }
        lock_version: { type: integer, minimum: 1 }
...
    CreditIncreaseRequest:
      type: object
      required: [id, folio, distributor_id, requested_amount, status, origin, line_snapshot, reasons, actors, requested_at, lock_version]
      properties:
        id: { type: string, format: uuid }
        folio: { type: string }
        distributor_id: { oneOf: [{ type: string, format: uuid }, { type: integer }] }
        branch_id: { type: integer }
        coordinator_id: { type: integer }
        requested_amount: { $ref: '#/components/schemas/Money' }
        recommended_amount: { oneOf: [{ $ref: '#/components/schemas/Money' }, { type: 'null' }] }
        authorized_amount: { oneOf: [{ $ref: '#/components/schemas/Money' }, { type: 'null' }] }
        origin: { type: object }
        line_snapshot: { type: object }
        status: { $ref: '#/components/schemas/CreditIncreaseStatus' }
        reasons: { type: object }
        actors: { type: object }
        requested_at: { type: string, format: date-time }
        reviewed_at: { type: [string, 'null'], format: date-time }
        decided_at: { type: [string, 'null'], format: date-time }
        restriction_id: { type: [string, integer, 'null'] }
        lock_version: { type: integer }
    DomainError:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, correlationId]
          properties:
            code:
              type: string
              enum: [AUTH_SCOPE_DENIED, REAUTHENTICATION_REQUIRED, RESOURCE_VERSION_CONFLICT, CREDIT_LINE_ALREADY_EXISTS, CREDIT_LINE_NOT_FOUND, CREDIT_INVALID_BALANCE, CREDIT_INSUFFICIENT, CREDIT_50_PERCENT_RULE_NOT_SATISFIED, CREDIT_50_PERCENT_NO_ADMISSIBLE_AMOUNT, CREDIT_RESTRICTION_ALREADY_BOUND, CREDIT_RESTRICTION_NOT_ACTIVE, CREDIT_RESTRICTION_VOUCHER_MISMATCH, CREDIT_INCREASE_INVALID_STATE, CREDIT_INCREASE_AMOUNT_INVALID, CREDIT_INCREASE_NOT_PREAUTHORIZED, CREDIT_INCREASE_EXCEEDS_REQUESTED, CREDIT_MOVEMENT_DUPLICATE, CAPITAL_RECOVERY_EXCEEDS_USED_BALANCE]
            message: { type: string }
            correlationId: { type: string, format: uuid }
...